refactor(core): split pdf.rs into coherent private modules
CORE-T12 maintainability refactor: replace the oversized conv/pdf.rs with a
pdf/ module tree while preserving to_markdown/detailed diagnostics behavior,
PDFium TLS cache + global call/init mutexes, and pdf-extract panic containment.

Private boundaries:
- pdfium: native binding cache and call serialization
- inspector: structured pdf-inspector fast/parallel/OCR-aware paths
- ocr: render/page-image OCR helpers
- fallback: PDFium char-count path and pdf-extract
- native_text: native layer extraction and trust/coverage gates
- postprocess: malformed-table detection and margin stripping
- recovery: needs_ocr recovery classification and warning sources

Unit tests move with their owning logic. No heuristic, warning-code, or
dependency changes.
